requirements.txt
Django~=4.2.8
djangorestframework~=3.14.0